lib/AI/requirements.txt
absl-py @ file:///croot/absl-py_1714140470852/work
aiohttp @ file:///croot/aiohttp_1715108769648/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
astor==0.8.1
astunparse==1.6.3
async-timeout @ file:///croot/async-timeout_1703096998144/work
attrs @ file:///croot/attrs_1695717823297/work
blinker @ file:///croot/blinker_1696539051175/work
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools @ file:///croot/cachetools_1713977085444/work
certifi @ file:///croot/certifi_1720453481653/work/certifi
cffi @ file:///croot/cffi_1714483155441/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click @ file:///croot/click_1698129812380/work
cryptography @ file:///croot/cryptography_1714660666131/work
Flask==3.0.3
flatbuffers @ file:///croot/python-flatbuffers_1722369094438/work
frozenlist @ file:///croot/frozenlist_1698702560391/work
gast @ file:///Users/ktietz/demo/mc3/conda-bld/gast_1628588903283/work
google-auth @ file:///croot/google-auth_1715111149956/work
google-auth-oauthlib @ file:///opt/conda/conda-bld/google-auth-oauthlib_1660687784486/work
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
grpcio @ file:///croot/grpc-split_1716834572340/work
h5py @ file:///tmp/build/80754af9/h5py_1608649014695/work
idna @ file:///croot/idna_1714398848350/work
importlib-metadata @ file:///croot/importlib_metadata-suite_1704813515092/work
itsdangerous==2.2.0
Jinja2==3.1.4
Keras @ file:///Users/ktietz/demo/mc3/conda-bld/keras_split_1630579936814/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
Markdown @ file:///croot/markdown_1671541909495/work
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
multidict @ file:///croot/multidict_1701096859099/work
numpy @ file:///work/mkl/numpy_and_numpy_base_1682952636358/work
oauthlib @ file:///croot/oauthlib_1679489621486/work
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
pillow==11.0.0
protobuf==4.25.3
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyJWT @ file:///croot/pyjwt_1715094751094/work
pyOpenSSL @ file:///croot/pyopenssl_1708380408460/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
PyYAML @ file:///croot/pyyaml_1698096049011/work
requests @ file:///croot/requests_1721410876868/work
requests-oauthlib @ file:///croot/requests-oauthlib_1720615017878/work
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scipy @ file:///croot/scipy_1717521478074/work/dist/scipy-1.13.1-cp39-cp39-linux_x86_64.whl#sha256=2cb3d1436f8762a53161486cb9493ccc305fc760e8f830f82888792cbaf99099
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard @ file:///home/builder/mesters/opt/envs/tensorflow/conda-bld/tensorboard_1682445826165/work/tensorboard-2.12.1-py3-none-any.whl
tensorboard-data-server @ file:///croot/tensorboard-data-server_1681498183723/work/tensorboard_data_server-0.7.0-py3-none-manylinux2014_x86_64.whl
tensorboard-plugin-wit @ file:///home/builder/tkoch/workspace/tensorflow/tensorboard-plugin-wit_1658918494740/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorflow @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_cpu/tensorflow-base_1614583966145/work/tensorflow_pkg/tensorflow-2.4.1-cp39-cp39-linux_x86_64.whl
tensorflow-estimator @ file:///home/builder/adipietro/tf/tensorflow-estimator_1630508970172/work/tensorflow_estimator-2.6.0-py2.py3-none-any.whl
termcolor @ file:///croot/termcolor_1668084651543/work
urllib3 @ file:///croot/urllib3_1718912636303/work
Werkzeug @ file:///croot/werkzeug_1716533313107/work
wrapt @ file:///tmp/abs_c335821b-6e43-4504-9816-b1a52d3d3e1eel6uae8l/croots/recipe/wrapt_1657814400492/work
yarl @ file:///croot/yarl_1701105127787/work
zipp @ file:///croot/zipp_1704206909481/work
